Move each format into a separate part file.
